+++ CHANGES.rst
@@ -5,7 +5,11 @@
 3.1.2 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Fix the psycopg2cffi driver inadvertently depending on the
+  ``psycopg2`` package. See :issue:`403`.
+- Make the error messages for unavailable drivers include more
+  information on underlying causes.
+- Log a debug message when an "auto" driver is successfully resolved.
 
 
 3.1.1 (2020-07-02)